Created separate files and fixed README linting
