+++ utils/characters.py
@@ -1,0 +1,5 @@
+def wordCount(string):
+    return len(string.split())
+
+def charCount(string):
+    return len(string)
